build: use WITH_XXX instead of XXX_FOUND
To check if something should be en- or disabled WITH_XXX
should be used and not XXX_FOUND.

If XXX_FOUND is used and something gets disabled afterwards (by setting
WITH_XXX to OFF) it will be compiled in as long as XXX_FOUND is found in
cmake's cache file. So disabling a feature, or option, without
clearing the CMakeCache.txt might result in builds with unwanted
configuration.