libfreerdp-utils/pcap: open pcap in binary mode
Use 'b' in fopen's mode string to force binary (untranslated) mode
when reading or writing the pcap file.
This is required on WIN32 and maybe on other non POSIX conforming
systems.

This fixes the mstsc internal error disconnect followed by a crash
of the sample server when latter is used to serve a pcap file on
WIN32.